Skip if caption already exists to save time

+++ image_captioning/im2txt/run_inference.py
@@ -82,6 +82,11 @@ def worker(image_urls):
 
       url = url.strip()
 
+      # Check if caption already exists. If so, ignore
+      out_filename = url.split('/')[-1] + '.txt'
+      if os.path.exists(os.path.join(CAPTIONS_DIR, out_filename)):
+        continue
+
       img_filename = 'image_tmp_%d' % idx
       try:
         urllib.urlretrieve(url, img_filename)
@@ -132,11 +137,11 @@ def main(_):
   with open(FLAGS.input_image_urls, "r") as f:
     image_urls = f.readlines()
   
-  chunk_size = int(math.ceil(len(image_urls) / float(mp.cpu_count())))
+  chunk_size = 10000
   print("Chunk Size: %d" % chunk_size)
  
   chunks = [(i, image_urls[x:x+chunk_size]) for i, x in enumerate(xrange(0, len(image_urls), chunk_size))]
-  pool = mp.Pool(processes=mp.cpu_count())
+  pool = mp.Pool(processes=mp.cpu_count() - 1)
   pool.map(worker, chunks)
   pool.close()
